+++ test/goalc/test_with_game.cpp
@@ -771,6 +771,10 @@ TEST_F(WithGameTests, SoundName) {
                           "0\n"});
 }
 
+TEST_F(WithGameTests, StaticLambda) {
+  runner.run_static_test(env, testCategory, "test-static-lambda.gc", {"Add: 30 sub: -10\n0\n"});
+}
+
 TEST(TypeConsistency, TypeConsistency) {
   Compiler compiler;
   compiler.enable_throw_on_redefines();